+++ tests/test.php
@@ -29,6 +29,7 @@ include_once("boletoHolmes.php");
 include_once("boletoHolmesLog.php");
 include_once("invoice.php");
 include_once("invoiceLog.php");
+include_once("dictKey.php");
 include_once("transfer.php");
 include_once("transferLog.php");
 include_once("boletoPayment.php");